feat: ch06 상품 수정

+++ src/main/java/com/kosa/shop/repository/ItemImgRepository.java
@@ -5,9 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ItemImgRepository extends JpaRepository<ItemImg, Integer> {
+public interface ItemImgRepository extends JpaRepository<ItemImg, Long> {
 
-//    List<ItemImg> findByItemId(Long itemId);
     List<ItemImg> findByItemIdOrderById(Long itemId);
 
 }
